Added basic JS for task 1
Today's struggle: could not get the JS running for a while. It seemed undetected even though everything was correctly linked. After some trial and error as well as research, I identified the missing element (the monthDisplay in JS was not in HTML). I added it to my HTML as a span and it worked.

I now have the basic structure and task 1 is done.